requirements.txt
python-telegram-bot==20.7
deep-translator==1.11.4
python-dotenv==1.0.0
pymongo==4.6.1 